Adding undo

+++ app/src/main/java/com/cmu/banavision/CameraViewModel.kt
@@ -6,6 +6,7 @@ import android.content.pm.PackageManager
 import android.graphics.Bitmap
 import android.location.Geocoder
 import android.net.Uri
+import android.util.Log
 import android.widget.Toast
 import androidx.core.app.ActivityCompat
 import androidx.lifecycle.ViewModel
@@ -15,6 +16,7 @@ import com.cmu.banavision.util.LocationState
 import com.google.android.gms.location.FusedLocationProviderClient
 import com.google.android.gms.location.LocationServices
 import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.update
@@ -34,25 +36,39 @@ class CameraViewModel @Inject constructor(
     val imageUris = _imageUris.asStateFlow()
     private val _locationState = MutableStateFlow(LocationState())
     val locationState = _locationState.asStateFlow()
+    private val _pendingDeleteImage = MutableStateFlow<Uri?>(null)
+    val pendingDeleteImage = _pendingDeleteImage.asStateFlow()
 
-    fun deleteImage(uri: Uri,context: Context){
+    fun deleteImage(uri: Uri, context: Context) {
         viewModelScope.launch {
+            _pendingDeleteImage.value = uri
             _imageUris.update { imageState ->
                 imageState.copy(
                     uris = imageState.uris - uri
                 )
             }
-           val deletedImage= pictureUseCase.captureAndSaveImageUseCase.deleteImage(context,uri)
-            if(deletedImage){
-                print("Image deleted")
+            // delete image in 5 seconds when undo is not clicked
+            delay(5000)
+            if (pendingDeleteImage.value == uri) {
+                pictureUseCase.captureAndSaveImageUseCase.deleteImage(context, uri)
+                _pendingDeleteImage.value = null
             }
         }
 
     }
+ fun undoDeleteImage(uri: Uri) {
+    viewModelScope.launch {
+        _pendingDeleteImage.value = null
+        _imageUris.update { imageState ->
+            imageState.copy(
+                uris = imageState.uris + uri
+            )
+        }
+    }
+}
 
     fun onTakePhoto(uri: Uri, context: Context) {
         viewModelScope.launch {
-            // convert bitmap to uri
             _imageUris.update { imageState ->
                 imageState.copy(
                     uris = imageState.uris + uri
@@ -60,6 +76,7 @@ class CameraViewModel @Inject constructor(
             }
             getLocation(context)
         }
+
     }
 
     fun chooseImageFromGallery(uri: Uri, context: Context) {
@@ -71,20 +88,33 @@ class CameraViewModel @Inject constructor(
             }
             getLocation(context)
         }
+
     }
 
     private fun getLocation(context: Context) {
-        if (ActivityCompat.checkSelfPermission(context, Manifest.permission.ACCESS_FINE_LOCATION) != PackageManager.PERMISSION_GRANTED &&
-            ActivityCompat.checkSelfPermission(context, Manifest.permission.ACCESS_COARSE_LOCATION) != PackageManager.PERMISSION_GRANTED) {
+        if (ActivityCompat.checkSelfPermission(
+                context,
+                Manifest.permission.ACCESS_FINE_LOCATION
+            ) != PackageManager.PERMISSION_GRANTED &&
+            ActivityCompat.checkSelfPermission(
+                context,
+                Manifest.permission.ACCESS_COARSE_LOCATION
+            ) != PackageManager.PERMISSION_GRANTED
+        ) {
             // Request location permissions
             ActivityCompat.requestPermissions(
                 context as MainActivity,
-                arrayOf(Manifest.permission.ACCESS_FINE_LOCATION, Manifest.permission.ACCESS_COARSE_LOCATION),
+                arrayOf(
+                    Manifest.permission.ACCESS_FINE_LOCATION,
+                    Manifest.permission.ACCESS_COARSE_LOCATION
+                ),
                 0
             )
 
         } else {
-            val fusedLocationClient: FusedLocationProviderClient = LocationServices.getFusedLocationProviderClient(context)
+           Log.i( "Location", "Location permission granted")
+            val fusedLocationClient: FusedLocationProviderClient =
+                LocationServices.getFusedLocationProviderClient(context)
             fusedLocationClient.lastLocation.addOnSuccessListener { location ->
                 if (location != null) {
                     val latitude = location.latitude
@@ -101,23 +131,32 @@ class CameraViewModel @Inject constructor(
                         longitude = longitude,
                         locationName = locationName
                     )
-
+                }
+                else {
+                    Log.i("Location", "Location is null")
                 }
             }
         }
     }
 
-   fun bitmapToUri(context: Context, it: Bitmap): Uri {
-           return pictureUseCase.captureAndSaveImageUseCase.bitmapToUri(context, it)
+    fun bitmapToUri(context: Context, it: Bitmap): Uri {
+        return pictureUseCase.captureAndSaveImageUseCase.bitmapToUri(context, it)
     }
 
-     fun clearImageUris() {
-         viewModelScope.launch {
-             _imageUris.update { imageState ->
-                 imageState.copy(
-                     uris = emptyList()
-                 )
-             }
-         }
-     }
+    fun clearImageUris(context: Context) {
+        viewModelScope.launch {
+            _imageUris.value.uris.forEach {
+                if (it != null) {
+                    pictureUseCase.captureAndSaveImageUseCase.deleteImage(context, it)
+                    _pendingDeleteImage.value = it
+                }
+            }
+            _imageUris.update { imageState ->
+                imageState.copy(
+                    uris = emptyList()
+                )
+            }
+        }
+
+    }
 }
